feat(domain): implementa camada de domínio com DDD e design patterns
Domain Model:
- Entities: Media, WatchEntry, User, Person, Credit, Season, Episode, Genre
- Value Objects: SeriesProgress, Rating, Duration
- Enums: MediaType, WatchStatus, CreditType

Domain Events (PR1):
- WatchEntryCreatedEvent, WatchEntryUpdatedEvent
- MediaCreatedEvent, MediaUpdatedEvent
- Event base class: DomainEvent (timestamp, aggregateId)

Design Patterns Implementados:

1. Template Method (PR5):
- AbstractMediaValidator: validate() template
- MovieValidator: validações específicas de filmes
- SeriesValidator: validações específicas de séries

2. Factory Method (PR5):
- MediaValidatorFactory: getValidator(type)
- Singleton pattern para validators

3. Strategy Pattern (PR6):
- RecommendationStrategy interface
- ContentBasedRecommendationStrategy
- CollaborativeRecommendationStrategy
- HybridRecommendationStrategy

4. Builder Pattern:
- DomainException.Builder
- Event builders

Error Handling:
- DomainException: exceções específicas do domínio
- ErrorCode enum
- Validation errors

Value Objects:
- SeriesProgress: season, episode, watchedDuration, totalDuration
- Immutability garantida
- Equals/hashCode baseados em valor

Validations:
- Título obrigatório (1-500 chars)
- Ano: 1888-atual+5 (filmes), 1950-atual+3 (séries)
- Overview: max 5000 chars
- Rating: 0.0-10.0
- Template Method: validações customizadas por tipo

PR1: Event Envelope & Infrastructure
PR5: Design Patterns Integration
PR6: Recommendations System
ADR-001: Hexagonal Architecture
ADR-008: Design Patterns Implementation

Refs: #domain #ddd #patterns

+++ src/main/java/com/cine/cinelog/core/domain/policy/RatingPolicy.java
@@ -1,9 +1,20 @@
 package com.cine.cinelog.core.domain.policy;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 import com.cine.cinelog.core.domain.model.WatchEntry;
+/**
+ * Política de domínio para gerenciamento de rating.
+ * Define as regras e validações relacionadas a rating.
+ * 
+ * <p>Esta política encapsula lógica de negócio específica
+ * e é aplicada durante operações em Rating.</p>
+ * 
+ * @since 1.0
+ * @see Rating
+ */
 
 public interface RatingPolicy {
-    void validateCanRate(WatchEntry entry, Integer rating, Instant when);
+    void validateCanRate(WatchEntry entry, BigDecimal rating, Instant when);
 }
